+++ lib/src/madari_engine_base.dart
@@ -8,9 +8,15 @@ import 'package:madari_engine/src/services/list_service.dart';
 import 'package:madari_engine/src/services/profile_service.dart';
 import 'package:pocketbase/pocketbase.dart';
 
-@JS('MadariEngine')
+export './services/addon_service/types.dart';
+export './services/catalog_service/types.dart';
+export './services/list_service/index.dart';
+export './services/profile_service/types.dart';
+
+@JSExport()
 class MadariEngine {
-  final PocketBase pb;
+  @JSExport()
+  late final PocketBase _pb;
   late final AuthService authService;
   late final AddonService addonService;
   late final CatalogService catalogService;
@@ -19,8 +25,9 @@ class MadariEngine {
   late final ListService listService;
 
   MadariEngine({
-    required this.pb,
+    required PocketBase pb,
   }) {
+    _pb = pb;
     authService = AuthService(pb: pb);
     authService.setup();
     addonService = AddonService(pb: pb);
@@ -40,5 +47,14 @@ class MadariEngine {
     profileService.setup();
     listService = ListService(pb: pb, profileService: profileService);
     listService.setup();
+  }
+
+  dispose() {
+    authService.dispose();
+    addonService.dispose();
+    catalogService.dispose();
+    exploreService.dispose();
+    profileService.dispose();
+    listService.dispose();
   }
 }