+++ lib/src/services/addon_service/types.dart
@@ -1,0 +1,48 @@
+import 'dart:js_interop';
+
+import 'package:pocketbase/pocketbase.dart';
+
+class PocketBaseStremioAddon {
+  final String? id;
+  final String url;
+  final String title;
+  final String? icon;
+  final bool enabled;
+
+  PocketBaseStremioAddon({
+    required this.url,
+    required this.title,
+    required this.icon,
+    required this.enabled,
+    this.id,
+  });
+
+  @JSExport()
+  PocketBaseStremioAddon copyWith({
+    String? url,
+    String? title,
+    String? icon,
+    bool? enabled,
+    String? id,
+  }) {
+    return PocketBaseStremioAddon(
+      url: url ?? this.url,
+      title: title ?? this.title,
+      icon: icon ?? this.icon,
+      enabled: enabled ?? this.enabled,
+      id: id ?? this.id,
+    );
+  }
+
+  @JSExport()
+  Map<String, dynamic> toJson(PocketBase pb) {
+    return {
+      if (id != null) 'id': id,
+      'url': url,
+      'title': title,
+      'icon': icon,
+      'enabled': enabled,
+      'user': pb.authStore.record!.id,
+    };
+  }
+}