+++ lib/src/services/list_service/index.dart
@@ -1,7 +1,9 @@
 import 'dart:convert';
+import 'dart:js_interop';
 
 import 'package:madari_engine/src/services/profile_service.dart';
 
+@JSExport()
 class ListModel {
   final String id;
   final String name;
@@ -41,6 +43,7 @@ class ListModel {
     };
   }
 
+  @JSExport()
   ListModel copyWith({
     String? name,
     String? description,
@@ -59,6 +62,7 @@ class ListModel {
   }
 }
 
+@JSExport()
 class ListItemModel {
   final String id;
   final String type;
@@ -116,6 +120,7 @@ class CreateListRequest {
     required this.description,
   });
 
+  @JSExport()
   Future<Map<String, dynamic>> toJson(ProfileService service) async {
     return {
       'name': name,
@@ -137,6 +142,7 @@ class UpdateListRequest {
     required this.description,
   });
 
+  @JSExport()
   Map<String, dynamic> toJson() {
     return {
       'name': name,