+++ lib/web.dart
@@ -1,6 +1,14 @@
-@JS()
-library madari_engine;
-
 import 'dart:js_interop';
 
-void main() {}
+import 'package:pocketbase/pocketbase.dart';
+
+@JS('setupMadari')
+external set _setupMadari(JSFunction f);
+
+void setupMadari(JSString baseURL) {
+  final pb = PocketBase(baseURL.toDart);
+}
+
+void main() {
+  _setupMadari = setupMadari.toJS;
+}